Rename Math to Nat
Nat is slightly more descriptive, as it accurately describes that
we provide predicates indexed by Nat. It also is compatible with
a hypothetical future where we can index by other numeric types
in a world with dependent types.

We also change some names and implementations in the name of
ergonomics / simplicity.

For starters, we:
  - Shorten Equals to Eq.
  - Shorten inequality names.

We also change inequality implementations to refer to each other,
e.g., LTE = LT or Eq. This falls under the 'simplicity' category,
as the error messages are arguably worse. But it simplifies the
API and implementation, so it seems worth it.